[ENH]: error helpers & common WrappedSqlxError type
