+++ rust/error/Cargo.toml
@@ -7,4 +7,10 @@ edition = "2021"
 path = "src/lib.rs"
 
 [dependencies]
-tonic = { workspace = true }
+tonic = { workspace = true, optional = true }
+sqlx = { workspace = true, optional = true }
+thiserror = { workspace = true }
+
+[features]
+tonic = ["dep:tonic"]
+sqlx = ["dep:sqlx"]
