+++ rust/error/src/lib.rs
@@ -3,7 +3,17 @@
 // Custom errors can use these codes in order to allow for generic handling
 use std::error::Error;
 
-#[derive(PartialEq, Debug)]
+#[cfg(feature = "tonic")]
+mod tonic;
+#[cfg(feature = "tonic")]
+pub use tonic::*;
+
+#[cfg(feature = "sqlx")]
+mod sqlx;
+#[cfg(feature = "sqlx")]
+pub use sqlx::*;
+
+#[derive(PartialEq, Debug, Clone, Copy)]
 pub enum ErrorCodes {
     // OK is returned on success, we use "Success" since Ok is a keyword in Rust.
     Success = 0,
@@ -45,6 +55,12 @@ pub enum ErrorCodes {
 
 pub trait ChromaError: Error + Send {
     fn code(&self) -> ErrorCodes;
+    fn boxed(self) -> Box<dyn ChromaError>
+    where
+        Self: Sized + 'static,
+    {
+        Box::new(self)
+    }
 }
 
 impl Error for Box<dyn ChromaError> {}
@@ -52,54 +68,5 @@ impl Error for Box<dyn ChromaError> {}
 impl ChromaError for Box<dyn ChromaError> {
     fn code(&self) -> ErrorCodes {
         self.as_ref().code()
-    }
-}
-
-impl From<ErrorCodes> for tonic::Code {
-    fn from(err: ErrorCodes) -> tonic::Code {
-        match err {
-            ErrorCodes::Success => tonic::Code::Ok,
-            ErrorCodes::Cancelled => tonic::Code::Cancelled,
-            ErrorCodes::Unknown => tonic::Code::Unknown,
-            ErrorCodes::InvalidArgument => tonic::Code::InvalidArgument,
-            ErrorCodes::DeadlineExceeded => tonic::Code::DeadlineExceeded,
-            ErrorCodes::NotFound => tonic::Code::NotFound,
-            ErrorCodes::AlreadyExists => tonic::Code::AlreadyExists,
-            ErrorCodes::PermissionDenied => tonic::Code::PermissionDenied,
-            ErrorCodes::ResourceExhausted => tonic::Code::ResourceExhausted,
-            ErrorCodes::FailedPrecondition => tonic::Code::FailedPrecondition,
-            ErrorCodes::Aborted => tonic::Code::Aborted,
-            ErrorCodes::OutOfRange => tonic::Code::OutOfRange,
-            ErrorCodes::Unimplemented => tonic::Code::Unimplemented,
-            ErrorCodes::Internal => tonic::Code::Internal,
-            ErrorCodes::Unavailable => tonic::Code::Unavailable,
-            ErrorCodes::DataLoss => tonic::Code::DataLoss,
-            ErrorCodes::Unauthenticated => tonic::Code::Unauthenticated,
-            ErrorCodes::VersionMismatch => tonic::Code::Internal,
-        }
-    }
-}
-
-impl From<tonic::Code> for ErrorCodes {
-    fn from(code: tonic::Code) -> ErrorCodes {
-        match code {
-            tonic::Code::Ok => ErrorCodes::Success,
-            tonic::Code::Cancelled => ErrorCodes::Cancelled,
-            tonic::Code::Unknown => ErrorCodes::Unknown,
-            tonic::Code::InvalidArgument => ErrorCodes::InvalidArgument,
-            tonic::Code::DeadlineExceeded => ErrorCodes::DeadlineExceeded,
-            tonic::Code::NotFound => ErrorCodes::NotFound,
-            tonic::Code::AlreadyExists => ErrorCodes::AlreadyExists,
-            tonic::Code::PermissionDenied => ErrorCodes::PermissionDenied,
-            tonic::Code::ResourceExhausted => ErrorCodes::ResourceExhausted,
-            tonic::Code::FailedPrecondition => ErrorCodes::FailedPrecondition,
-            tonic::Code::Aborted => ErrorCodes::Aborted,
-            tonic::Code::OutOfRange => ErrorCodes::OutOfRange,
-            tonic::Code::Unimplemented => ErrorCodes::Unimplemented,
-            tonic::Code::Internal => ErrorCodes::Internal,
-            tonic::Code::Unavailable => ErrorCodes::Unavailable,
-            tonic::Code::DataLoss => ErrorCodes::DataLoss,
-            tonic::Code::Unauthenticated => ErrorCodes::Unauthenticated,
-        }
     }
 }
