+++ rust/sysdb/Cargo.toml
@@ -27,7 +27,7 @@ tower = { workspace = true }
 sqlx = { workspace = true }
 
 chroma-config = { workspace = true }
-chroma-error = { workspace = true }
+chroma-error = { workspace = true, features = ["tonic"] }
 chroma-types = { workspace = true }
 chroma-tracing = { workspace = true, features = ["grpc"] }
 chroma-sqlite = { workspace = true }
