+++ rust/types/Cargo.toml
@@ -22,7 +22,7 @@ pyo3 = { workspace = true }
 proptest = { workspace = true, optional = true }
 proptest-derive = { workspace = true, optional = true }
 
-chroma-error = { workspace = true }
+chroma-error = { workspace = true, features = ["tonic"] }
 chroma-config = { workspace = true }
 
 [build-dependencies]
